% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{.findBlocks}
\alias{.findBlocks}
\title{Internal for identify blocks in a sequence}
\usage{
.findBlocks(x, thr = 0.03)
}
\arguments{
\item{x}{A vector of integers}

\item{thr}{The threshold to identify blocks, as fraction of total length}
}
\value{
A vector with the blocks labeled by the dominant integer in the block
}
\description{
Internal for identify blocks in a sequence
}
