% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_estimateTM.R
\name{estimateTM}
\alias{estimateTM}
\title{Estimate transition matrix for SMC model components}
\usage{
estimateTM(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Estimate transition matrix for SMC model components
}
