% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-multinomial.R
\name{ExactMultinomialTest}
\alias{ExactMultinomialTest}
\title{Exact Multinomial Test}
\usage{
ExactMultinomialTest(observed, prob)
}
\arguments{
\item{observed}{Vector (or matrix) of observed values. The length
(or number of columns) defines the number of categories of the multinomial
distribution. The sum of the vector (or rows) the sample size.}

\item{prob}{The theoretical probability of the multinomial distribution.}
}
\description{
Exact Multinomial Test
}
