% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-multinomial.R
\name{multinomial_support}
\alias{multinomial_support}
\title{Support of a multinomial distribution}
\usage{
multinomial_support(k, n)
}
\arguments{
\item{k}{The number of categories.}

\item{n}{The sample size.}
}
\description{
Support of a multinomial distribution
}
