% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-distributions.R
\name{rbrokenbar}
\alias{rbrokenbar}
\title{Broken bar distribution}
\usage{
rbrokenbar(n, S, sequential = FALSE, shuffle = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{S}{number of parts to break the unit bar. See details.}

\item{sequential}{Boolean. If TRUE, the small pieces is sequentially broken. See details.}
}
\description{
Broken bar distribution
}
