% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-distributions.R
\name{rdunif}
\alias{rdunif}
\title{Title}
\usage{
rdunif(n, min = 0, max = 1)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{min}{lower limit of the distribution. Must be integer and finite.}

\item{max}{upper limit of the distribution. Must be integer and finite.}
}
\description{
Title
}
