% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-sampling.R
\name{sampling}
\alias{sampling}
\title{Sampling simulator}
\usage{
sampling(object, skip = 0, start = 0, fraction = 1, ...)
}
\arguments{
\item{object}{Population to be sampled, usually a data.frame}

\item{skip}{Number of individuals to skip before next observation.
The default is zero.}

\item{start}{Start of the sampling as a fraction of the total observations.
The default is zero (sampling start at the beginning).}

\item{fraction}{Fraction to be sampled (after start). The default is one,
all the population is sampled.}

\item{...}{Additional arguments to be passed to other methods.}
}
\value{
The sample from the population.
}
\description{
Sampling simulator
}
