% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkbox-smc.R
\name{smcSim}
\alias{smcSim}
\title{Simulate a transition matrix for a Sequential Markov Chaing}
\usage{
smcSim(G, S, L)
}
\arguments{
\item{N}{}
}
\description{
Simulate a transition matrix for a Sequential Markov Chaing
}
